requirements.txt
Flask==2.3.2
pandas==2.1.1
beautifulsoup4==4.12.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
torch==2.2.0